requirements.txt
flask==2.3.3
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
tensorflow==2.16.1
flask-wtf==1.1.1
python-dotenv==1.0.0
werkzeug==2.3.7
jinja2==3.1.2
flask-assets==2.0
gunicorn==21.2.0
pytz==2023.3
cssmin==0.2.0
jsmin==3.0.1
flask-session==0.5.0
yfinance==0.2.35
zoneinfo==0.2.1;python_version<"3.9" 